+++ runtime/service/src/main/java/org/apache/polaris/service/catalog/iceberg/IcebergCatalog.java
@@ -804,7 +804,7 @@ public class IcebergCatalog extends BaseMetastoreViewCatalog
     boolean purge =
         callContext
             .getRealmConfig()
-            .getConfig(FeatureConfiguration.PURGE_VIEWS_ON_DROP, catalogEntity);
+            .getConfig(FeatureConfiguration.PURGE_VIEW_METADATA_ON_DROP, catalogEntity);
 
     return dropTableLike(PolarisEntitySubType.ICEBERG_VIEW, identifier, Map.of(), purge)
         .isSuccess();