+++ CHANGELOG.md
@@ -50,7 +50,7 @@ at locations that better optimize for object storage.
 - The Helm chart now supports Pod Disruption Budgets (PDBs) for Polaris components. This allows users to define
   the minimum number of pods that must be available during voluntary disruptions, such as node maintenance.
 
-- Feature configuration `PURGE_VIEWS_ON_DROP` was added to allow dropping views without purging their metadata files.
+- Feature configuration `PURGE_VIEW_METADATA_ON_DROP` was added to allow dropping views without purging their metadata files.
 
 ### Changes